Fix circular import error
