+++ ramp/core/constants.py
@@ -1,3 +1,25 @@
+def switch_on_parameters():
+    """
+    Calibration parameters. These can be changed in case the user has some real data against which the model can be calibrated
+    They regulate the probability of coincident switch-on within the peak window
+
+    mu_peak corresponds to \mu_{%} in [1], p.8
+    s_peak corresponds to \sigma_{%} in [1], p.8
+
+    Notes
+    -----
+    [1] F. Lombardi, S. Balderrama, S. Quoilin, E. Colombo,
+        Generating high-resolution multi-energy load profiles for remote areas with an open-source stochastic model,
+        Energy, 2019, https://doi.org/10.1016/j.energy.2019.04.097.
+    """
+
+    mu_peak = 0.5  # median value of gaussian distribution [0,1] by which the number of coincident switch_ons is randomly selected
+    s_peak = 0.5  # standard deviation (as percentage of the median value) of the gaussian distribution [0,1] above mentioned
+    op_factor = 0.5  # off-peak coincidence calculation parameter
+
+    return mu_peak, s_peak, op_factor
+
+
 OLD_TO_NEW_MAPPING = {
     "name": "user_name",
     "n_users": "num_users",
